common: 64 bits intrinsic fixes
long is 8B on linux. Replace it with s32
long long is always 8B but let's replace it with s64 so we are sure

+++ common/include/Utilities/General.h
@@ -111,7 +111,7 @@ protected:
 class BaseDeletableObject : public virtual IDeletableObject
 {
 protected:
-	volatile long	m_IsBeingDeleted;
+	volatile s32	m_IsBeingDeleted;
 
 public:
 	BaseDeletableObject();
